+++ docs/CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Add google analytics
 - Add MLH 2020 badge
 
 ## [1.3.0](https://github.com/hackmcgill/mchacks7/tree/1.3.0) - 2019-12-17